Squashed commit of the following:
commit a48e3d5909ea9b1c3ee62b3cd5a405dc692ebc92
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Tue Oct 5 12:13:03 2021 +0800

    Add implementation for `AppchainLifecycleManager`.

commit 2be8c7d623afc740975fb6a3cb5b6713f8653583
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Tue Oct 5 11:19:14 2021 +0800

    Add implementation of `AnchorViewer`.

commit 9a71b358a1ec1a9e4b25d1e1d66c44608d8ffd1c
Author: Rivers Yang <rivers.yang@icloud.com>
Date:   Tue Oct 5 10:25:58 2021 +0800

    Add implementation for `ft_on_transfer`.

+++ appchain-anchor/src/types.rs
@@ -70,7 +70,7 @@ pub struct ProtocolSettings {
     pub maximum_market_value_percent_of_wrapped_appchain_token: u16,
     /// The minimum number of validator(s) registered in this contract for
     /// booting the corresponding appchain and keep it alive.
-    pub minimum_validator_count: u16,
+    pub minimum_validator_count: u64,
     /// The maximum number of validator(s) which a delegator can delegate to.
     pub maximum_validators_per_delegator: u64,
     /// The unlock period (in days) for validator(s) can withdraw their deposit after
